WIP catch NotAuthorizedError when processing entity event updates
This should handle the case where a user has been removed from a group for which they have event updates. Needs more investigation though, we still need to figure out how to best reproduce cases where this can happen.

It seems that the server sends a 404 when trying to receive CalendarEvents for a calendar group that has been deleted, rather than 403. While this is still not ideal (we should detect this before passing off to the main thread) the gui isn't ending up in a weird broken state. So it's probably a different case?