WIP enable sender name for email aliases. #516
We enable sender name for aliases. In this version, if sending from an alias that doesn't have a sender name, it chooses the sender name of the account. It could be that we want to leave it empty in this case?

Also, I still need to handle sender names in the AddEmailAddressesPage of AddDomainWizard