Invalidate the usage assignments cache on error 400, p#26
When we run into an error 400 (BadRequest), then
the test is likely to have been deleted from tutadb
and the cached assignments are out of date.

We simply invalidate the cache and log a message
so that they are reloaded on the next login.

Attempting to retry sending the ping is likely futile here
as a BadRequestError would occur if there is a mismatch in
configured and programmed test or if the test has been deleted.